glossary_terms: move 2 forms to Views/
Per #4331:
- Components::GlossaryTermForm
    -> Views::Controllers::GlossaryTerms::Form
- Components::ImagesToRemoveForm
    -> Views::Controllers::GlossaryTerms::Images::RemoveForm

ImagesToRemoveForm lives under the nested glossary_terms/images/
subtree because the only caller is glossary_terms/images_controller.
Three ERB callers updated.
