+++ app/views/controllers/glossary_terms/edit.html.erb
@@ -3,4 +3,4 @@ add_edit_title(@glossary_term)
 add_context_nav(glossary_term_form_edit_tabs(term: @glossary_term))
 %>
 
-<%= render(Components::GlossaryTermForm.new(@glossary_term)) %>
+<%= render(Views::Controllers::GlossaryTerms::Form.new(@glossary_term)) %>
