+++ app/views/controllers/glossary_terms/new.html.erb
@@ -3,7 +3,7 @@ add_new_title(:create_object, :GLOSSARY_TERM)
 add_context_nav(glossary_term_form_new_tabs)
 %>
 
-<%= render(Components::GlossaryTermForm.new(
+<%= render(Views::Controllers::GlossaryTerms::Form.new(
   @glossary_term,
   enctype: "multipart/form-data",
   upload_params: {
